Move Node version guard to CLI runtime utilities
Relocated assertSupportedNodeVersion from src/cli/manual/manual-cli-helpers.js to src/cli/runtime/node-version.js so CLI entry points can enforce supported Node versions without loading manual-only progress bar and HTTP helpers.
