+++ src/cli/generate-gml-identifiers.js
@@ -6,8 +6,8 @@ import vm from "node:vm";
 import { Command, InvalidArgumentError } from "commander";
 
 import { CliUsageError, handleCliError } from "./cli-errors.js";
+import { assertSupportedNodeVersion } from "./runtime/node-version.js";
 import {
-    assertSupportedNodeVersion,
     createManualGitHubClient,
     ensureDir,
     formatDuration,
